Modify map zoom to keep the point under the mouse cursor constant
The idea is that while zooming the map the point under the mouse would
stay as close to constant as possible (given that we use integer
coordinates).

This version uses some algebra to figure out the correct new parameters
for the mercator projection used in osm-gps-map.

Occasionally (and we haven't figured out what triggers it) zooming out
suddenly resets your position to 0,0.

[Dirk Hohndel: switched this to using the correct interface to get the
               object properties and did some serious whitespace cleanup]

Signed-off-by: Robert C. Helling <helling@atdotde.de>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>