Don't mark divelist changed for automated merges after all
The intention was good, and this was added in commit 4982389ca762 ("Fix
setting of the dive_table.preexisting logic"), but it turns out to not
be that great idea after all.

So the thinking is that merging two dives clearly changes the dive list,
and it really does.  At the same time, because it's an automated merge,
if you re-read the old XML file, you'll get it done again, so saving the
changes doesn't really *matter*.

And it turns out to be somewhat annoying with test dives: we have

 - dives/test23.xml:
    <dive number='23' tripflag='INTRIP' date='2011-12-02' time='6:00:00' duration='30:00 min'>
 - dives/test25.xml:
    <dive number='26' date='2011-12-02' time='6:00:00' duration='30:00 min'>

that merge automatically if you run subsurface on all the test dives
together.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>