UTF8 aware parser for some more GPS formats
I'm sure there are better ways to do this, but this appears to grok most
rational formats I was able to find. NSEW or positive/negative numbers.
Decimal degrees (WGS84) or degrees and decimal minutes (that's what most
GPSs seem to provide). I'm sure there are still corner cases that confuse
it, but it seemed reasonably robust in testing.

I don't really love the ';' as separator but that solves the obvious
problem with locales that use a decimal comma.

Signed-off-by: Dirk Hohndel <dirk@hohndel.org>